#292 refactor: 패키지 가독성을 위해 도메인 오브젝트 폴더 분류
