+++ backend/src/main/java/moheng/recommendtrip/domain/preferredlocation/PreferredLocationsFinder.java
@@ -1,0 +1,7 @@
+package moheng.recommendtrip.domain.preferredlocation;
+
+import java.util.Map;
+
+public interface PreferredLocationsFinder {
+    Map<Long, Long> findPreferredLocations(final long memberId);
+}
