+++ backend/src/main/java/moheng/trip/domain/model/ExternalSimilarTripModelClient.java
@@ -1,7 +1,6 @@
-package moheng.trip.domain;
+package moheng.trip.domain.model;
 
 import moheng.trip.dto.FindSimilarTripWithContentIdResponses;
-import moheng.trip.dto.SimilarTripResponses;
 
 public interface ExternalSimilarTripModelClient {
     FindSimilarTripWithContentIdResponses findSimilarTrips(final long contentId, final long page);
