+++ backend/src/test/java/moheng/config/stub/StubKeywordFilterModelClient.java
@@ -2,7 +2,7 @@ package moheng.config.stub;
 
 import moheng.keyword.dto.TripContentIdsByKeywordResponse;
 import moheng.keyword.dto.TripRecommendByKeywordRequest;
-import moheng.keyword.domain.KeywordFilterModelClient;
+import moheng.keyword.domain.model.KeywordFilterModelClient;
 
 import java.util.List;
